test(netcdf): cover scan metadata range paths
